Make clearer separation of sessions and allow customized implementations
