+++ mars/deploy/oscar/ray.py
@@ -38,18 +38,18 @@ from ...services.cluster.backends.base import (
 )
 from ...services import NodeRole
 from ...services.task.execution.api import ExecutionConfig
+from ...session import (
+    _new_session,
+    new_session,
+    AbstractSession,
+    ensure_isolation_created,
+)
 from ...utils import lazy_import, retry_callable
 from ..utils import (
     load_config,
     get_third_party_modules_from_config,
 )
 from .service import start_supervisor, start_worker, stop_supervisor, stop_worker
-from .session import (
-    _new_session,
-    new_session,
-    AbstractSession,
-    ensure_isolation_created,
-)
 from .pool import create_supervisor_actor_pool, create_worker_actor_pool
 
 ray = lazy_import("ray")
