+++ mars/deploy/oscar/tests/test_local.py
@@ -44,10 +44,7 @@ from ....lib.aio import new_isolation
 from ....storage import StorageLevel
 from ....services.storage import StorageAPI
 from ....services.task.supervisor.task import TaskProcessor
-from ....tensor.arithmetic.add import TensorAdd
-from ....tests.core import mock, check_dict_structure_same, DICT_NOT_EMPTY
-from ..local import new_cluster, _load_config
-from ..session import (
+from ....session import (
     get_default_async_session,
     get_default_session,
     clear_default_session,
@@ -60,9 +57,12 @@ from ..session import (
     ExecutionInfo,
     Profiling,
     Progress,
-    _IsolatedWebSession,
     _execute_with_progress,
 )
+from ....tensor.arithmetic.add import TensorAdd
+from ....tests.core import mock, check_dict_structure_same, DICT_NOT_EMPTY
+from ..local import new_cluster, _load_config
+from ..session import _IsolatedWebSession
 from ..tests.session import new_test_session
 from .modules.utils import (  # noqa: F401; pylint: disable=unused-variable
     cleanup_third_party_modules_output,
