+++ mars/deploy/oscar/tests/test_ray_dag.py
@@ -20,11 +20,11 @@ import pytest
 
 from .... import get_context
 from .... import tensor as mt
+from ....session import new_session, get_default_async_session
 from ....tests import test_session
 from ....tests.core import DICT_NOT_EMPTY, require_ray
 from ....utils import lazy_import
 from ..local import new_cluster
-from ..session import new_session, get_default_async_session
 from ..tests import test_local
 from ..tests.session import new_test_session
 from ..tests.test_local import _cancel_when_tile, _cancel_when_execute
