+++ mars/deploy/oscar/tests/test_ray_load_modules.py
@@ -18,13 +18,13 @@ import numpy as np
 import pytest
 
 from .... import tensor as mt
+from ....session import get_default_session, new_session
 from ....tests.core import require_ray
 from ....utils import lazy_import
 from ..ray import (
     _load_config,
     new_cluster,
 )
-from ..session import get_default_session, new_session
 from .modules.utils import (  # noqa: F401  # pylint: disable=unused-variable
     cleanup_third_party_modules_output,
     get_output_filenames,
