+++ mars/learn/contrib/tsfresh/core.py
@@ -15,7 +15,7 @@
 import pandas as pd
 
 from .... import remote as mr
-from ....deploy.oscar.session import get_default_session
+from ....session import get_default_session
 from ....utils import ceildiv
 
 try:
