+++ mars/remote/tests/test_run_script.py
@@ -33,7 +33,7 @@ assert session is not None
 
 script3 = b"""
 from mars.core.operand import Fetch
-from mars.deploy.oscar.session import AbstractSession
+from mars.session import AbstractSession
 
 assert AbstractSession.default is not None
 assert isinstance(tensor.op, Fetch)
