+++ setup.cfg
@@ -193,5 +193,5 @@ exclude =
     mars/lib/version.py
 
 [codespell]
-ignore-words-list = hist,rcall,fpr,ser,nd,inout,ot,Ba,ba,asend,hart,coo,splitted,datas,fro
+ignore-words-list = hist,rcall,fpr,ser,nd,inout,ot,Ba,ba,asend,hart,coo,splitted,datas,fro,rabit,ccompiler
 skip = .idea,.git,./build,./docs/build,./mars/lib,node_modules,static,generated,*.po,*.ts,*.json,*.c,*.cpp,*.cfg
